Moved live-query folder up
