+++ src/classes/dexie/dexie-static-props.ts
@@ -17,8 +17,8 @@ import { errnames } from '../../errors';
 import { getMaxKey } from '../../functions/quirks';
 import { vip } from './vip';
 import { globalEvents } from '../../globals/global-events';
-import { liveQuery } from '../live-query/live-query';
-import { extendObservabilitySet } from '../live-query/extend-observability-set';
+import { liveQuery } from '../../live-query/live-query';
+import { extendObservabilitySet } from '../../live-query/extend-observability-set';
 
 /* (Dexie) is an instance of DexieConstructor, as defined in public/types/dexie-constructor.d.ts
 *  (new Dexie()) is an instance of Dexie, as defined in public/types/dexie.d.ts
